Add article deletion Popup shows up when trying to create a list without budget Update budget progress bar in ListCompositionScreen

+++ app/src/main/java/ca/uqac/etu/jcid/chadal/ui/ShoppingListViewModel.kt
@@ -30,8 +30,8 @@ class ShoppingListViewModel : ViewModel() {
         }
     }
 
-    private fun calculateTotal(): Float {
-        var total:Float = 0f
+    private fun calculateTotal(): Double {
+        var total:Double = 0.0
         _uiState.value.articles.forEach {
             // Tax calculation & application
             total += it.price * (1 + it.category.taxPercentage)
